fix: Fix area specific death counts being always zero
The API returning death counts does not fill in the "healthCareDistrict"
field, but instead fills in "area" field which seems to have hospital
names in it. Since korona-info shows district-specific death counts, we
would need to map the hospitals (or other more granular entities
reported by the API) to healthcare districts. This commit implements a
simple lookup table for hospitals and health care districts as well as a
simple lookup method for the data in chartDataHelper.ts

Closes #62